requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0
redis==5.0.1
celery==5.3.4
yt-dlp==2023.11.16
openai-whisper==20231117
openai==1.3.7
moviepy==1.0.3
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1 